Prepend view path so mustache take precedence.
This change is driven by the need to override devise erb views with mustache
views and templates. Because the view path is appended, the views in the gem get
detected first and used as a result of the way the view paths are traversed. By
prepending the view_path, the mustache views are found first, which if you've
taken the time to put anything here is almost certianly what you want.